reports/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
Jinja2==3.1.4
httpx==0.27.2
PyJWT==2.9.0
